renamed tokenize to maskSensitiveInfo
