+++ apps/etl-func/src/adapters/functions/__tests__/messages-ingestion.test.ts
@@ -6,7 +6,7 @@ import {
 import { messageSchema } from "@/adapters/avro.js";
 import { MessageAdapter, MessageContentProvider } from "@/adapters/message.js";
 import { EventHubEventProducer } from "@/adapters/message-event.js";
-import PDVTokenizerClient from "@/adapters/pdv-tokenizer/pdv-tokenizer-client.js";
+import PDVTokenizerClient from "@/adapters/tokenizer/pdv-tokenizer-client.js";
 import { InvocationContext } from "@azure/functions";
 import { pino } from "pino";
 import { afterEach, describe, expect, test, vi } from "vitest";
@@ -44,7 +44,7 @@ const getMessageByMetadataSpy = vi
 
 const PDVTokenizer = new PDVTokenizerClient("testApiKey", "testBaseUri");
 const tokenizeSpy = vi
-  .spyOn(PDVTokenizer, "tokenize")
+  .spyOn(PDVTokenizer, "maskSensitiveInfo")
   .mockResolvedValue("9314a1ea-ac0b-11ef-9cd2-0242ac120002");
 
 const eventHubProducerClient = new mocks.EventHubProducerClient();
